refactoring
- grouped code into seperate libs
