+++ src/sys_stats/mod.rs
@@ -1,0 +1,5 @@
+pub mod cpu;
+pub mod disk;
+pub mod memory;
+pub mod socket;
+pub mod sys_stats_handler;
